refactor: move hack/cert-generator to cmd/cert-generator
