+++ pkg/version/VERSION
@@ -1,1 +1,1 @@
-v0.4.7-debug.1
+v0.4.7-debug.2
